Initial commit
This is a simple script that downloads a Guix binary.

My end goal is to build an archive with Guix, that contains a Brotli-
enabled Nginx, and an Acme client for Letsencrypt. I will run that under
systemd on CoreOS.

In order to build the archive, I need a working Guix package manager.
One way to obtain it is by using GuixSD, but as I don't use GuixSD as my
normal machine, I would have to run it inside a virtual machine, which
is a bit unfortunate. Especially if I want to automate building, that
involves automating a virtual machine, which is tedious. For e.g.
running on Travis CI, having a simple script that installs Guix would be
easier.

Now of course I don't want to install Guix on my normal machine (yet),
as it is a bit invasive. So I am installing it in an Ubuntu 16.04
container which is running on systemd-nspawn. The situation is quite
ridiculous. But in the end I hope this script can run as-is on Travis,
and if you want to run it locally, that can be done in a systemd-nspawn
container.

Also add a .travis.yml so Travis can build the same binaries. Then we
can check that the builds are bit by bit reproducible.

Some notes:

Write down steps to build in systemd-nspawn. I just had to give it
the right capability to make the builds work in my local container. Now
I have an Ubuntu 16.04 container transformed into something that runs
Guix. Yay!

Make the place where the GuixSD commit hash is stored into a dedicated
file. This makes it easier to track changes that are just rolling the
version.

Collect garbage after a build to help to keep the build machine tidy.
Set minimum free space, rather than min to collect. 512 MiB free space
should be sufficient to do a subsequent build. If we always GC a minimum
amount, that might GC things that are actually useful, so if there is no
need to GC then don't.