Give in to the shouty all caps convention
I have long avoided this in my repositories, because I think it is
impolite to shout at the users. But there is something to say for the
argument that it makes the readme easier to spot in a list of files. And
as there is no hierarchy in this repository, it might make sense to put
the readme in the conventional shouty all-caps location.