Add license
License everything under the GNU General Public License, version 3
specifically. (And not any later version.)

I am not a lawyer and this is not legal advice. I do not know whether
overriding a package definition that is part of GNU Guix constitutes a
"modification" (in which case I must license the "entire work, as a
whole", whatever that is, under the GNU GPL), or if it constitutes a
"compilation of a covered work with other separate and independent
works". To sidestep the entire issue, I will just license all code in
this repository under the GNU General Public License as well.