Ensure /usr/sbin is in the path in build.sh
For some reason, when I ran the build script again in a fresh container,
suddenly /usr/sbin was not in the path any more. Even though the Ubuntu
image I use was still the same ... I don't really understand why this
happened, but for now let's work around by explicitly adding /usr/sbin
to the path.