Add Nix installation script
Building with Guix is failing sometimes with obscure error messages. It
also times out on Travis, and updating is very very slow for Guix. (It
takes a few minutes to "load packages", even on my machine.)

The last time I looked at Nix it was badly documented, very spammy, and
full of magic incantations. But now that I am considering it again, the
documentation seems decent (commands have good man pages), there is a
new 'nix' command with a more intuitive user interface. It is no longer
spammy. And above all, it is fast. There is also the advantage that Nix
appears to be more widely used than Guix, so it probably is more up to
date. It has stable channels that receive security updates. So let's
give Nix another try.

I can build a stock Nginx and acme-client. Next step is to figure out
how to build Nginx with Brotli support, and then how to generate an
image from all of this.