Build a squashfs image using Nix
This is far easier than I would have thought -- once you find out how to
do it by trial and error and inspecting random gists on Github ... There
is a lack of documentation, but the tools work nicely.

So now it produces a squashfs image (better than a tar, because systemd
can mount this as a RootImage=). Unfortunately there is some cruft in
the image that I do not want, but let's clear that up as a next step.