Enable ngx_brotli module
I have to admit, Nix is really good. Doing this in Guix was much more
convoluted than in Nix. Here it is really, define the source, add the
configure flag, done. No need to copy-paste the entire Nginx package. No
need to define a fake package for sources only, as support for that is
built in. Really neat.

Now the only issue with this might be the same one as on Guix: by
passing the store path of the ngx_brotli module, Nix will detect a
dynamic dependency on it, because Nginx embeds the configure command
line. Let's fix that next.