Remove Guix-related files
Bye Guix, it was good to know you. I like how you forced me to finally
learn Scheme a bit better. I was amazed by the quality of your
documentation, and I found the idea of Scheme for everything elegant.

But after giving Nix another try, it was easier to accomplish my goal in
the end. (I have a working squashfs image now. With Guix I only got to a
tarball.) The build is also much faster, and it requires less scripting
to drive. Pinning to a specific Nixpkgs version can be reified, unlike
Guix where I had to externally pull a specific revision and then do the
build. With Nix, different versions of Nixpkgs can live alongside each
other, just like any package can. And though the Nix syntax looked
alien to me two days ago, after studying it a bit and reading the Nix
Pills, it makes a lot of sense, and I find it quite elegant too.

So long Guix, and thanks for all the fish.