Remove apparmor and audit dependencies
Apparmor is not useful on the minimal system that I am building: we can
use systemd's isolation features already which are more powerful. (I
assume they don't depend on libapparmor though.) And even if there is an
exploit, there is nothing on the server to exploit; that is the idea of
a minimal server.

As for audit, I have never used it and it is unclear to me what it does.
Let's remove it to reduce the closure size. Again, security features are
more important if you want to isolate different workloads on the same
system. But I intend to run only Nginx on this system (and *only* that).