Build a rootfs for minimal server
Okay, so I have a file system with the binaries I need. I can even use
systemd-nspawn --boot to make *something* happen, but then it exits with
"Parent died too early". Probably because my entire filesystem is empty,
and there are no targets to start?

Anyway, this is an interesting starting point. It is quite bloated
still even after removing as much packages as I could. 227 MiB for just
systemd and a few small applications! About 170 of those are taken up by
systemd itself. I should see if there is a way to reduce that.