Use nixos-unstable channel as base
I submitted a pull request to Nixpkgs to update LibreSSL, as there was a
new release for 2.6 and 2.7 with security fixes. There was no update for
2.5 however, which is in NixOS 18.03. I am not sure at this point if
that is going to be backported. I think it will be easier for me to stay
up to date with security fixes if work from the unstable channel (the
lastes commit on master that built succesfully). "Unstable" sounds scary
but I think it will not be an issue in practice, especially if I do my
own QA, which I should do anyway.