Restore reproducibility (configure /etc/group bug)
The Nginx configure script reads the NGX_USER from /etc/group if none is
set explicitly, and this causes the build output to differ depending on
whether a "nobody" group exists on the host system. Specify the group
explicitly to avoid this issue. I now get the same nginx binary whether
I build on Arch Linux or inside an Ubuntu 16.04 container.