Include sha256 of pinned Nixpkgs archive
This prevents having to fetch it again and again even if the url did not
change. Because we assume that the archive at that url is stable, we can
provide a hash, and then there is no need to download the given Nixpkgs
revision if the commit hash did not change.

This also adds a 'nix-prefetch-url' invocation to the update script.
It will prefetch the url and store the hash in the pinned Nixpkgs
expression. What is stored in nixpkgs-pinned.nix is no longer a string
literal, but a full fetchTarball Nix expression.