Begin writing acme-client unit
It at least shows a help screen now. I'm not sure how I will get the
domains in there though. The primary domain is easy enough with a
systemd template, but what about the alternate ones? Store them in an
EnvironmentFile that depends on the %i unit argument? Feels a bit hacky,
but I don't know of a much better way either.