Start working on h2o version
H2O looks nice. The package in Nixpkgs has too many dependencies though,
there is an almost 4 MB binary size regression, and we now have a
dependency on Bash, which is very bad. But I think this is not
essential, it will just require a bit of pruning.

And then it would require porting all of my carefully crafted Nginx
config to h2o. Not sure if it is worth it at this point. The two
advantages I can see are support for dynamic TLS record sizes, and
built-in support for static Brotli-compressed files. But the latter I
now have in Miniserver, and the former might not help *that* much.