Revert container madness
I can't get anything to run under rkt. The documentation is outdated and
contradicts the implementation. The appc format is deprecated, but the
successor "oci" not yet supported, it seems. Runc only wants directory
trees and it expects some kind of config json file that I am not sure
where it should come from, but Skopeo definitely did not produce it.
What is all of this madness, I'll go back to systemd then. It is just as
broken, but at least I can do with one mountable squashfs and a single
service file, rather than having to deal with manifests and registries
and whatnot. </rant>

This reverts commit a419abdbba6baeed025255c645962889bad93ec5.
This reverts commit b11527595f0983eb70d1fe5e86c8b9ce3561a4b9.