Revert "Compile LibreSSL with -z noexecstack"
It should no longer be necessary, now that I upstreamed a fix in
Nixpkgs, which will be included with the next Nixpkgs update.

There is no risk of accidentally keeping executable stacks enabled, at
least not with the current systemd config, as Nginx will fail to load
libcrypto if it requires executable stacks.

This reverts commit 46b8ec174f2317805a58b9ec28d0e21a82491a9d.