Upgrade to Nixpkgs with squashfs-tools 4.4
I am preparing a pull request to upgrade squashfs-tools. The
reproducibility patches have been included upstream. To verify that
images are reproducible, I will try to build Miniserver locally and on
CI (which has a different number of cores, and which will run at a
different time). If the output differs, we know that the the patches are
still required.