Remove trailing comma in Github workflow config
I use json to avoid shooting myself in the foot with yaml. I introduce a
bug in the json that makes it not parse as json. But it is still valid
yaml, so the build did not break. Not sure if this is a good thing or a
bad thing.