Recover from Nix upstream installer deletion
The installer is no longer hosted at the url where it used to be, and
that broke CI.

I don't think the install script is still needed, we download and
verify the archive directly, which contains the real install script.
For single-user installation we also don't need to set up build users.

This also changes the version to 2.3.4.