Write GitHub build config as yaml after all
It is more readable than the json, the caveats of yaml are when writing,
and yaml is easier to edit than json (trailing comma issues and such).

I did keep all of the strings quoted, hopefully that is a good middle
ground between avoiding yaml footguns and having a config that is easy
on the eye.