Extract group from package name
Sometimes Nix bumps the Perl or Python version, but not the version of
the individual Perl packages. But because the Perl version shows up in
their name, it cause them to show up as a big pile of additions and
removals. Ignore these changes instead, based on a hard-coded list of
package groups known to do this.

When we do it this way, the differ does not need to be changed, because
the differ does not look at groups. So if the group changes but the
packages stay unchanged, those packages will not be reported.

I might still want to print the names in the formatter though.