Support exact revisions in Nix update script
I update the pinned revision once in a while (not just here, for all my
projects), and then I end up with fifty copies of slightly different
versions of Nixpkgs on my machine. Instead of always pinning to the
latest revision, I could also pin to a recent one that I used before, to
avoid some churn. Support that, by accepting a commit hash in addition
to a branch name.