Generate reproducible dm-verity data
This adds a layer of security on top of the read-only squashfs image:
tampering with the image itself would be detected. Arguably this is not
very useful, because presumably if you could modify the image, you could
also modify the verity file. But it does create an additional barrier
for attackers. Also, it protects against bitrot in the image file, when
the underlying file system doesn't already do that.