Update acme-client, nginx, and 19 changes
This updates the pinned Nixpkgs snapshot to the latest commit in the
nixos-unstable branch of nixos/nixpkgs.

Runtime dependencies:

  acme-client 1.2.0  -> 1.3.0
  nginx       1.21.6 -> 1.22.0
+ openssl               3.0.3  (!)
- pcre        8.45
+ pcre2                 10.40

Build dependencies:

  cargo           1.60.0  -> 1.61.0
  cargo-bootstrap 1.59.0  -> 1.60.0
  coreutils       9.0     -> 9.1
  docbook-xml     4.5     -> 4.3
+ docbook-xml                4.5
  git-minimal     2.36.0  -> 2.36.1
+ gtk-doc                    1.33.2
  lvm2            2.03.15 -> 2.03.16
  mercurial       6.1.2   -> 6.1.3
  ncurses         6.3     -> 6.3-p20220507
+ pcre                       8.45
- pcre2           10.39
  python3         3.9.12  -> 3.9.13
  python3-minimal 3.9.12  -> 3.9.13
  rustc           1.60.0  -> 1.61.0
  rustc-bootstrap 1.59.0  -> 1.60.0

So this is not right, we gained a runtime dependency on OpenSSL,
that should be LibreSSL instead. Let's investigate and fix this in
a follow-up commit.

Also, the switch between pcre and pcre2 in build time vs runtime is
interesting, not sure what was going on there.