Replace acme-client with Lego
The error messages that acme-client generates are abysmal. It looks
like it's incompatible with the v02 API. Let's go with this Go program
instead.

The binary size does grow by a *lot*, unfortunately. I guess that
happens when Lego is a static executable and it brings its own TLS stack
implementation, instead of me being able to use LibreSSL. For now I will
accept this, hopefully something can be done about it in the future.