Add unit to reload nginx config
After Lego refreshes the config, we need to send a SIGHUP to nginx to
make it pick up the new certificate. The easiest way to do this, is to
do it unconditionally after Lego runs, regardless of whether it actually
refreshed the certificate.

We do this in a separate unit, because in the Lego unit itself, there is
the RootImage=, so it doesn't have systemctl available (even when we
prefix the command with '+', that only affects users, not RootImage=).