Update libressl to 3.8.3, and 21 changes
This updates the pinned Nixpkgs snapshot to the latest commit in the
nixos-unstable branch of nixos/nixpkgs.

Runtime dependencies:

  libressl 3.8.2 -> 3.8.3

Build dependencies:

  build                      1.0.3      -> 1.1.1
  buildcatrust               0.2.1      -> 0.3.0
  cargo-auditable            0.6.1      -> 0.6.2
  cython                     3.0.7      -> 3.0.9
  freezegun                  1.3.1      -> 1.4.0
  git                        2.43.2     -> 2.44.0
  git-minimal                2.43.2     -> 2.44.0
  go                         1.21.7     -> 1.22.1
  lzip                       1.24       -> 1.24.1
  nghttp2                    1.59.0     -> 1.60.0
  pathspec                   0.11.2     -> 0.12.1
  pytest                     7.4.4      -> 8.0.2
  pytest-asyncio             0.23.5     -> 0.23.5.post1
  python-dateutil            2.8.2      -> 2.9.0.post0
  python3.11-bootstrap-build 1.0.3      -> 1.1.1
  re2                        2024-02-01 -> 2024-03-01
  setuptools                 69.0.3     -> 69.1.1
  shadow                     4.14.5     -> 4.14.6
  trove-classifiers          2024.1.8   -> 2024.3.3
  typing-extensions          4.8.0      -> 4.10.0
  xz                         5.6.0      -> 5.6.1

Note, the xz version is the one that contains a backdoor: CVE-2024-3094.
NixOS is unaffected by all currently known exploit paths though, and for
Miniserver, this is not a runtime dependency, only a build-time
dependency, so unless the xz backdoor also hijacks the build of LibreSSL
or Nginx at build time, this should be acceptable ... though maybe to be
safe, I should downgrade xz at build time as well.